tests/wpt/web-platform-tests/svg/text/scripted/textpath-textlength-text-anchor-001.tentative.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:html="http://www.w3.org/1999/xhtml"><title>&apos;textLength&apos; on &lt;textPath&gt; with &apos;text-anchor&apos; adjustments</title><html:script src="/resources/testharness.js"/><html:script src="/resources/testharnessreport.js"/><html:link href="/fonts/ahem.css" rel="stylesheet"/><html:link href="https://svgwg.org/svg2-draft/text.html#TextPathElement" rel="help"/><html:link href="https://svgwg.org/svg2-draft/text.html#TextElementTextLengthAttribute" rel="help"/><defs><path id="p" d="M0,25h100"/></defs><g font-family="Ahem" font-size="20"><text expected-x="0"><textPath href="#p" lengthAdjust="spacing" text-anchor="start" textLength="25">XXXX</textPath></text><text expected-x="0" transform="translate(0 50)"><textPath href="#p" lengthAdjust="spacingAndGlyphs" text-anchor="start" textLength="25">XXXX</textPath></text><text expected-x="37.5"><textPath href="#p" lengthAdjust="spacing" startOffset="50%" text-anchor="middle" textLength="25">XXXX</textPath></text><text expected-x="37.5" transform="translate(0 50)"><textPath href="#p" lengthAdjust="spacingAndGlyphs" startOffset="50%" text-anchor="middle" textLength="25">XXXX</textPath></text><text expected-x="75"><textPath href="#p" lengthAdjust="spacing" startOffset="100%" text-anchor="end" textLength="25">XXXX</textPath></text><text expected-x="75" transform="translate(0 50)"><textPath href="#p" lengthAdjust="spacingAndGlyphs" startOffset="100%" text-anchor="end" textLength="25">XXXX</textPath></text></g><script>test(t =&gt; { let texts = Array.from(document.getElementsByTagName(&apos;text&apos;)); for (let text of texts) { let bbox = text.getBBox(); let textpath = text.firstElementChild; let description = &apos;text-anchor=&apos; + textpath.getAttribute(&apos;text-anchor&apos;); assert_approx_equals(bbox.x, parseFloat(text.getAttribute(&apos;expected-x&apos;)), 0.5, &apos;x, &apos; + description); assert_approx_equals(bbox.width, 25, 0.5, &apos;width, &apos; + description); } });</script></svg>
