LayoutTests/imported/w3c/web-platform-tests/svg/text/scripted/textpath-textlength-text-anchor-001.tentative-expected.txt

XXXX
XXXX
XXXX
XXXX
XXXX
XXXX

FAIL 'textLength' on <textPath> with 'text-anchor' adjustments assert_approx_equals: x, text-anchor=middle expected 37.5 +/- 0.5 but got 10
